backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.5.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
pytz==2024.1
